code review feedback - add shared ctypes
