+++ stdlib/3/multiprocessing/__init__.pyi
@@ -26,7 +26,6 @@ if sys.version_info >= (3, 8):
 if sys.platform != "win32":
     from multiprocessing.context import ForkContext, ForkServerContext
 
-
 # N.B. The functions below are generated at runtime by partially applying
 # multiprocessing.context.BaseContext's methods, so the two signatures should
 # be identical (modulo self).
@@ -50,39 +49,6 @@ def Pool(processes: Optional[int] = ...,
          initargs: Iterable[Any] = ...,
          maxtasksperchild: Optional[int] = ...) -> pool.Pool: ...
 
-class _Array:
-    value: Any = ...
-
-    def __init__(self, typecode_or_type: Union[str, Type[_CData]], size_or_initializer: Union[int, Sequence[Any]], *, lock: Union[bool, _LockLike] = ...) -> None: ...
-    def acquire(self) -> bool: ...
-    def release(self) -> bool: ...
-    def get_lock(self) -> _LockLike: ...
-    def get_obj(self) -> Any: ...
-
-    @overload
-    def __getitem__(self, key: int) -> Any: ...
-    @overload
-    def __getitem__(self, key: slice) -> List[Any]: ...
-    def __getslice__(self, start: int, stop: int) -> Any: ...
-    def __setitem__(self, key: int, value: Any) -> None: ...
-
-def Array(
-    typecode_or_type: Union[str, Type[_CData]],
-    size_or_initializer: Union[int, Sequence[Any]],
-    *,
-    lock: Union[bool, _LockLike] = ...) -> _Array: ...
-
-
-class _Value():
-    value: Any = ...
-    def __init__(self, typecode_or_type: Union[str, Type[_CData]], *args: Any, lock: Union[bool, _LockLike] = ...) -> None: ...
-    def get_lock(self) -> _LockLike: ...
-    def get_obj(self) -> Any: ...
-    def acquire(self) -> bool: ...
-    def release(self) -> bool: ...
-
-def Value(typecode_or_type: Union[str, Type[_CData]], *args: Any, lock: Union[bool, _LockLike] = ...) -> _Value: ...
-
 # ----- multiprocessing function stubs -----
 def allow_connection_pickling() -> None: ...
 def cpu_count() -> int: ...
@@ -94,6 +60,9 @@ def get_all_start_methods() -> List[str]: ...
 def get_start_method(allow_none: bool = ...) -> Optional[str]: ...
 def set_start_method(method: str, force: Optional[bool] = ...) -> None: ...
 
+_default_context: DefaultContext
+Array = DefaultContext.Array
+Value = DefaultContext.Value
 
 if sys.platform != "win32":
     @overload
