+++ stdlib/3/multiprocessing/context.pyi
@@ -1,5 +1,6 @@
 from logging import Logger
 import multiprocessing
+from multiprocessing import sharedctypes
 from multiprocessing import synchronize
 from multiprocessing import queues
 from multiprocessing.process import BaseProcess
@@ -79,7 +80,7 @@ class BaseContext(object):
         typecode_or_type: Any,
         *args: Any,
         lock: bool = ...
-    ) -> Any: ...
+    ) -> sharedctypes._Value: ...
     # TODO: typecode_or_type param is a ctype with a base class of _SimpleCData or array.typecode Need to figure out
     # how to handle the ctype
     # TODO: change return to Array once a stub exists in multiprocessing.sharedctypes
@@ -89,7 +90,7 @@ class BaseContext(object):
         size_or_initializer: Union[int, Sequence[Any]],
         *,
         lock: bool = ...
-    ) -> Any: ...
+    ) -> sharedctypes._Array: ...
     def freeze_support(self) -> None: ...
     def get_logger(self) -> Logger: ...
     def log_to_stderr(self, level: Optional[str] = ...) -> Logger: ...
